Minor adjustments.

+++ app/wakeit.qml
@@ -32,21 +32,23 @@ MainView {
         // It exists for backwards compatibility
         //======================================
         var db = LocalStorage.openDatabaseSync('WakeIt', '1.0', 'Wake It! Database', 10, function(db) {
-            // If the database doesn't exist, do nothing
+            db.changeVersion('', '1.0');
         });
         db.transaction(function(tx) {
-            var rs = tx.executeSql('SELECT host, mac FROM Device ORDER BY host');
-            for(var i = 0; i < rs.rows.length; ++i) {
-                deviceModel.add({
-                    title: "Untitled",
-                    local: false,
-                    host: rs.rows.item(i).host,
-                    mac: rs.rows.item(i).mac,
-                    port: 9
-                });
+            if (tx.executeSql('SELECT name FROM sqlite_master WHERE type="table" AND name="Device"').rows.length) {
+                var rs = tx.executeSql('SELECT host, mac FROM Device ORDER BY host');
+                for(var i = 0; i < rs.rows.length; ++i) {
+                    deviceModel.add({
+                        title: "Untitled",
+                        local: false,
+                        host: rs.rows.item(i).host,
+                        mac: rs.rows.item(i).mac,
+                        port: 9
+                    });
+                }
+                deviceModel.save();
+                tx.executeSql('DELETE FROM Device');
             }
-            deviceModel.save();
-            tx.executeSql('DELETE FROM Device');
         });
         //======================================
     }
